Reorganize dbAuth files
