+++ packages/api/src/functions/dbAuth/DbAuthHandler.js
@@ -1,8 +1,8 @@
 import CryptoJS from 'crypto-js'
 import { v4 as uuidv4 } from 'uuid'
 
-import * as DbAuthError from './dbAuthErrors'
-import { decryptSession, getSession } from './dbAuthHelpers'
+import * as DbAuthError from './errors'
+import { decryptSession, getSession } from './shared'
 
 export class DbAuthHandler {
   // class constant: maps the auth functions to their required HTTP verb for access
