+++ packages/api/src/functions/dbAuth/errors.ts
